Refactor the init process
